Added generated meta and TxFrame cleanup
